+++ modules/suppliers/local.php
@@ -10,7 +10,6 @@ require_once(__ROOT__.'/inc/settings.php');
 
 require_once(__ROOT__.'/func/getIngSupplier.php');
 require_once(__ROOT__.'/func/searchIFRA.php');
-//require_once(__ROOT__.'/func/getCatByID.php');
 require_once(__ROOT__.'/func/getDocument.php');
 
 $defCatClass = $settings['defCatClass'];
@@ -109,6 +108,7 @@ foreach ($ingredients as $ingredient) {
 	if($a = getIngSupplier($ingredient['id'],0,$conn)){ 
 		$j = 0;
 		unset($r['supplier']);
+		$err = null;
 		foreach ($a as $b){
 			$r['supplier'][$j]['name'] = (string)$b['name'];
 			$r['supplier'][$j]['link'] = (string)$b['supplierLink'];
@@ -117,6 +117,7 @@ foreach ($ingredients as $ingredient) {
 		}
 	}else{
 		$r['supplier'] = null;
+		$err = "No supplier is configured.\nYou won't be able to use this ingredient in a formula, unless at least one supplier is configured.";
 	}	
 	
 	if($d = getDocument($ingredient['id'],1,$conn)){
@@ -132,6 +133,7 @@ foreach ($ingredients as $ingredient) {
 	}
 	$r['stock'] = number_format((float)getIngSupplier($ingredient['id'],1,$conn)['stock'], $settings['qStep']) ?: 0;
 	
+	$r['error'] = $err;
 	$rx[]=$r;
 }